remove the grid module
